requirements.txt
beautifulsoup4
certifi
cffi
charset-normalizer
cryptography
cssutils
dnspython
frozendict
idna
loguru
lxml
more-itertools
numpy
pandas
pdfminer.six
pdfplumber
pillow
pycparser
pymongo
PyMuPDF
pypdfium2
pyrate-limiter
python-dateutil
python-dotenv
pytz
requests
sec-downloader
sec-edgar-downloader
sec-parser
six
soupsieve
tabulate
typing_extensions
tzdata
urllib3
xxhash
fastapi
uvicorn